requirements.txt
numpy==1.24.4
streamlit==1.2.0
Pillow==9.5.0
keras==2.10.0
tensorflow_intel==2.10.0
Keras-Preprocessing==1.1.2